# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "오픈마켓 API",
    "description": "오픈마켓 API Server입니다.<br>\n      <a href=\"/\">버전별 변경사항 확인</a>"
  },
  "servers": [
    {
      "url": "https://api.fesp.shop",
      "description": ""
    },
    {
      "url": "http://localhost",
      "description": "로컬 테스트"
    }
  ],
...
  "paths": {
    "/users/": {
      "post": {
        "tags": [
          "회원"
        ],
        "summary": "회원 가입",
        "description": "회원 가입을 합니다.<br>회원 가입을 완료한 후 회원 정보를 반환합니다.",
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/createUserRes"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/createUserResWithExtra"
                  }
                }
              }
            }
          },
          "409": {
            "description": "이메일 중복됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error409"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "requestBody": {
          "description": "회원 정보가 저장된 객체입니다.<br>email: 이메일(필수)<br>password: 비밀번호(필수)<br>name: 이름(필수)<br>phone: 전화번호(선택)<br>address: 주소(선택)<br>type: 회원 구분(필수, 구매회원: user, 판매회원: seller)<br>extra: 추가 데이터(선택). 추가하고 싶은 아무 속성이나 지정",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/createUser"
              },
              "examples": {
                "기본 속성": {
                  "$ref": "#/components/examples/createUser"
                },
                "extra 속성": {
                  "$ref": "#/components/examples/createUserWithExtra"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{_id}/{*}": {
      "get": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 조회(단일 속성)",
        "description": "회원 정보 중 한가지 속성을 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "description": "조회할 회원 id",
            "in": "path",
            "example": "5",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "*",
            "description": "조회할 속성<br>단일 속성 조회 예시: name<br>중첩 속성 조회 예시: extra/addressBook",
            "in": "path",
            "example": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "users/5/name 조회": {
                    "$ref": "#/components/examples/userInfoResOneAttr"
                  },
                  "users/5/extra/addressBook 조회": {
                    "$ref": "#/components/examples/userInfoResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
    "/users/{_id}": {
      "get": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 조회(모든 속성)",
        "description": "회원 정보의 모든 속성을 조회합니다.<br>\r  등록한 속성 이외에 다음의 정보가 추가됩니다.<br>\r  posts: 지정한 사용자가 작성한 게시글 수\r  postViews: 지정한 사용자가 작성한 모든 게시글 조회수\r  bookmark.products: 지정한 사용자가 북마크 한 상품 수\r  bookmark.users: 지정한 사용자가 북마크 한 사용자 수\r  bookmark.posts: 지정한 사용자가 북마크 한 게시글 수\r  ",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "조회할 회원 id",
            "example": "5"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/userInfoRes"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/userInfoResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 수정",
        "description": "회원 정보를 수정한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "회원 id",
            "example": "5"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/updateUserResOneAttr"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/updateUserResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "수정할 회원 정보",
          "required": true,
          "content": {
            "application/json": {
              "examples": {
                "기본 속성": {
                  "$ref": "#/components/examples/updateUserOneAttr"
                },
                "extra 속성": {
                  "$ref": "#/components/examples/updateUserWithExtra"
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/users/{_id}": {
      "get": {
        "tags": [
          "게시판"
        ],
        "summary": "사용자가 작성한 게시글 목록",
        "description": "사용자가 작성한 게시글 목록을 조회합니다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "사용자 id",
            "example": 4
          },
          {
            "name": "type",
            "description": "게시판 종류",
            "in": "query",
            "default": "post",
            "example": "qna",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "description": "검색어<br>제목과 내용 검색에 사용되는 키워드",
            "in": "query",
            "example": "배송",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "custom",
            "description": "custom 검색 조건",
            "in": "query",
            "example": "{'createdAt': {'$gte': '2024.04', '$lt': '2024.05'}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "description": "페이지",
            "in": "query",
            "example": 2,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "description": "한 페이지당 항목 수",
            "in": "query",
            "example": 10,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "description": "정렬(내림차순: -1, 오름차순: 1)",
            "in": "query",
            "example": "{'createdAt': 1}",
            "default": "{'createdAt': -1}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/postListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
...
    "/posts/{_id}/replies/{reply_id}": {
      "patch": {
        "tags": [
          "게시판"
        ],
        "summary": "댓글 수정",
        "description": "게시글의 댓글을 수정한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reply_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "게시판"
        ],
        "summary": "댓글 삭제",
        "description": "게시글의 댓글을 삭제한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reply_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Access Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      },
      "Refresh Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "error409": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "이미 등록된 리소스입니다."
          }
        }
      },
...
      "error500": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "요청하신 작업 처리에 실패했습니다. 잠시 후 다시 이용해 주시기 바랍니다."
          }
        }
      },
...
    },
    "examples": {
...
      "createUser": {
        "value": {
          "email": "gd@market.com",
          "password": "12345678",
          "name": "GD",
          "phone": "01011112222",
          "address": "서울시 강남구 역삼동 123",
          "type": "user",
          "profileImage": "/files/00-sample/profile.jpg"
        }
      },
...
      "createUserRes": {
        "value": {
          "ok": 1,
          "item": {
            "email": "gd@market.com",
            "name": "GD",
            "type": "user",
            "phone": "01011112222",
            "address": "서울시 강남구 역삼동 123",
            "_id": 5,
            "createdAt": "2023.11.21 14:31:15",
            "updatedAt": "2023.11.21 14:31:15"
          }
        }
      },
      "createUserResWithExtra": {
        "value": {
          "ok": 1,
          "item": {
            "email": "gdragon@market.com",
            "name": "G드래곤",
            "type": "user",
            "phone": "01011112222",
            "address": "서울시 강남구 역삼동 123",
            "extra": {
              "gender": "extra에는 프로젝트에서 필요한 아무 속성이나",
              "age": "넣으면 됩니다.",
              "address": [
                "배열도",
                "가능하고"
              ],
              "profileImage": {
                "path ": "/files/00-sample/profile.jpg",
                "name ": "profile.jpg",
                "originalname": "사커게임.png"
              },
              "obj": {
                "hello": "객체로",
                "hi": "넣어도 됩니다."
              },
              "addressBook": [
                {
                  "name": "집",
                  "address": "서울시"
                },
                {
                  "name": "회사",
                  "address": "인천시"
                }
              ]
            },
            "_id": 5,
            "createdAt": "2023.11.21 14:33:41",
            "updatedAt": "2023.11.21 14:33:41"
          }
        }
      },
...
    }
  }
}